Web.Payments.Cielo

Running the API calls

runCielo :: CieloConfig -> CieloM a -> IO a

API Calls happen in a MonadCielo type-class CieloM is a helper instance of this class, which runs exceptions on the IO monad and exposes the configuration through a ReaderT

cieloConfigFromEnv :: IO CieloConfig

We can load CieloConfig from the CIELO_MERCHANTID and CIELO_MERCHANTKEY environment variables

productionEnv :: Environment

sandboxEnv :: Environment

Available API Calls

Generating Merchant Order IDs

getMerchantOrderId :: MonadIO m => m Text

Generates a new merchant UIID

Sales

createSale :: MonadCielo m => Sale -> m Sale

Creates a Sale

querySale :: MonadCielo m => Text -> m Sale

Queries for a Sale given it's paymentId

captureSale

Arguments

:: MonadCielo m

=> Text
Payment ID

-> Maybe Int
Amount

-> Maybe Int
Service Tax Amount

-> m SaleUpdate

Captures a sale

voidSale

Arguments

:: MonadCielo m

=> Text
Payment ID

-> Maybe Int
Amount

-> m SaleUpdate

Voids a sale

Low-Level

updateSale

Arguments

:: MonadCielo m

=> Text
Payment ID

-> Text
:type

-> [(Text, Text)]
Data to PUT

-> m SaleUpdate

Updates a sale (Generalized PUT over a type of Sale update 1sales:id:type)

Querying Sales

querySalesByMerchantOrderId :: MonadCielo m => Text -> m SalesByMerchantOrderQuery

Queries for a sales given a merchant UUID

queryRecurrentSale :: MonadCielo m => Text -> m RecurrentPaymentQuery

Queries for a recurrent sale given it's paymentId

Recurrent Sales

Cancelling / Uncancelling

cancelRecurrentPayment :: MonadCielo m => Text -> m ()

uncancelRecurrentPayment :: MonadCielo m => Text -> m ()

Updates

updateRecurrentPaymentEndDate :: MonadCielo m => Text -> Text -> m ()

updateRecurrentPaymentPayment :: MonadCielo m => Text -> Payment -> m ()

updateRecurrentPaymentInterval :: MonadCielo m => Text -> Int -> m ()

updateRecurrentPaymentInstallments :: MonadCielo m => Text -> Int -> m ()

updateRecurrentPaymentCustomer :: MonadCielo m => Text -> Customer -> m ()

updateRecurrentPaymentRecurrencyDay :: MonadCielo m => Text -> Int -> m ()

updateRecurrentPaymentNextPaymentDate :: MonadCielo m => Text -> Text -> m ()

updateRecurrentPaymentNextPaymentAmount :: MonadCielo m => Text -> Int -> m ()

Types

data CieloConfig

Constructors

CieloConfig

Fields

cieloConfigMerchant :: Merchant

cieloConfigEnvironment :: Environment

type MonadCielo m = (MonadIO m, MonadReader CieloConfig m, MonadError CieloError m)

data CieloError

Constructors

CieloJSONError String SomeException ByteString

CieloInvalidError String HttpException (Maybe Value)

CieloNotFoundError String HttpException

CieloHttpException String HttpException (Maybe Value)

data Merchant

Constructors

Merchant

Fields

merchantId :: Text

merchantKey :: Text

data Environment

Constructors

Environment

Fields

environmentApiUrl :: Text

environmentApiQueryUrl :: Text

data Sale

Constructors

Sale

Fields

saleMerchantOrderId :: Text

saleCustomer :: Customer

salePayment :: Payment

data PaymentProvider

Constructors

PaymentProviderBradesco

PaymentProviderBancoDoBrasil

PaymentProviderSimulado

data PaymentType

Constructors

PaymentTypeCreditCard

PaymentTypeDebitCard

PaymentTypeElectronicTransfer

PaymentTypeBoleto

data Currency

Constructors

CurrencyBRL

CurrencyUSD

CurrencyMXN

CurrencyCOP

CurrencyCLP

CurrencyARS

CurrencyPEN

CurrencyEUR

CurrencyPYN

CurrencyUYU

CurrencyVEB

CurrencyVEF

CurrencyGBP

data Interval

Constructors

IntervalMonthly

IntervalBimonthly

IntervalQuarterly

IntervalSemiAnnual

IntervalAnnual

data RecurrentPayment

Constructors

RecurrentPayment

Fields

recurrentPaymentAuthorizeNow :: Maybe Bool

recurrentPaymentEndDate :: Maybe Text

recurrentPaymentStartDate :: Maybe Text

recurrentPaymentInterval :: Maybe Interval

recurrentPaymentRecurrentPaymentId :: Maybe Text

data RecurrentPaymentQuery

Constructors

RecurrentPaymentQuery

Fields

recurrentPaymentQueryRecurrentPayment :: RecurrentPayment

recurrentPaymentQueryCustomer :: Customer

data Address

Constructors

Address

Fields

addressStreet :: Maybe Text

addressNumber :: Maybe Text

addressComplement :: Maybe Text

addressZipCode :: Maybe Text

addressCity :: Maybe Text

addressState :: Maybe Text

addressCountry :: Maybe Text

data Customer

Constructors

Customer

Fields

customerName :: Text

customerEmail :: Maybe Text

customerBirthDate :: Maybe Text

customerIdentity :: Maybe Text

customerIdentityType :: Maybe Text

customerAddress :: Maybe Address

customerDeliveryAddress :: Maybe Address

data Payment

Constructors

Payment

Fields

paymentServiceTaxAmount :: Maybe Int

paymentInstallments :: Maybe Int

paymentInterest :: Maybe Value

Sometimes this is a string, sometimes it's an int... :P

paymentCreditCard :: Maybe CreditCard

paymentType :: Maybe PaymentType

paymentAmount :: Maybe Int

paymentCapture :: Maybe Bool

paymentAuthenticate :: Maybe Bool

paymentRecurrent :: Maybe Bool

paymentRecurrentPayment :: Maybe RecurrentPayment

paymentTid :: Maybe Text

paymentProofOfSale :: Maybe Text

paymentAuthorizationCode :: Maybe Text

paymentSoftDescriptor :: Maybe Text

paymentReturnUrl :: Maybe Text

paymentProvider :: Maybe PaymentProvider

paymentPaymentId :: Maybe Text

paymentReceivedDate :: Maybe Text

paymentCapturedAmount :: Maybe Int

paymentCapturedDate :: Maybe Text

paymentCurrency :: Maybe Currency

paymentCountry :: Maybe Text

paymentReturnCode :: Maybe Text

paymentReturnMessage :: Maybe Text

paymentStatus :: Maybe Int

paymentLinks :: Maybe [Value]

paymentExtraDataCollection :: Maybe [Value]

paymentExpirationDate :: Maybe Text

paymentUrl :: Maybe Text

paymentNumber :: Maybe Text

paymentBarCodeNumber :: Maybe Text

paymentDigitableLine :: Maybe Text

paymentAddress :: Maybe Text

data SalesByMerchantOrderQuery

Constructors

SalesByMerchantOrderQuery

Fields

salesByMerchantOrderQueryPayments :: [Payment]

Re-exports

class Default a where

A class for types with a default value.

Minimal complete definition

Nothing

class Monad m => MonadIO m where

Monads in which IO computations may be embedded. Any monad built by applying a sequence of monad transformers to the IO monad will be an instance of this class.

Instances should satisfy the following laws, which state that liftIO is a transformer of monads:

liftIO . return = return

liftIO (m >>= f) = liftIO m >>= (liftIO . f)

liftIO :: MonadIO m => forall a. IO a -> m a

Lift a computation from the IO monad.

class FromJSON a where

A type that can be converted from JSON, with the possibility of failure.

In many cases, you can get the compiler to generate parsing code for you (see below). To begin, let's cover writing an instance by hand.

There are various reasons a conversion could fail. For example, an Object could be missing a required key, an Array could be of the wrong size, or a value could be of an incompatible type.

The basic ways to signal a failed conversion are as follows:

empty and mzero work, but are terse and uninformative

fail yields a custom error message

typeMismatch produces an informative message for cases when the value encountered is not of the expected type

An example type and instance:

-- Allow ourselves to write Text literals.
{-# LANGUAGE OverloadedStrings #-}

data Coord = Coord { x :: Double, y :: Double }

instance FromJSON Coord where
  parseJSON (Object v) = Coord    <$>
                         v .: "x" <*>
                         v .: "y"

  -- We do not expect a non-Object value here.
  -- We could use mzero to fail, but typeMismatch
  -- gives a much more informative error message.
  parseJSON invalid    = typeMismatch "Coord" invalid

Instead of manually writing your FromJSON instance, there are two options to do it automatically:

Data.Aeson.TH provides Template Haskell functions which will derive an instance at compile time. The generated instance is optimized for your type so will probably be more efficient than the following two options:

The compiler can provide a default generic implementation for parseJSON.

To use the second, simply add a deriving Generic clause to your datatype and declare a FromJSON instance for your datatype without giving a definition for parseJSON.

For example, the previous example can be simplified to just:

{-# LANGUAGE DeriveGeneric #-}

import GHC.Generics

data Coord = Coord { x :: Double, y :: Double } deriving Generic

instance FromJSON Coord

If DefaultSignatures doesn't give exactly the results you want, you can customize the generic decoding with only a tiny amount of effort, using genericParseJSON with your preferred Options:

instance FromJSON Coord where
    parseJSON = genericParseJSON defaultOptions

Minimal complete definition

Nothing

class ToJSON a where

A type that can be converted to JSON.

An example type and instance:

-- Allow ourselves to write Text literals.
{-# LANGUAGE OverloadedStrings #-}

data Coord = Coord { x :: Double, y :: Double }

instance ToJSON Coord where
  toJSON (Coord x y) = object ["x" .= x, "y" .= y]

  toEncoding (Coord x y) = pairs ("x" .= x <> "y" .= y)

Instead of manually writing your ToJSON instance, there are two options to do it automatically:

Data.Aeson.TH provides Template Haskell functions which will derive an instance at compile time. The generated instance is optimized for your type so will probably be more efficient than the following two options:

The compiler can provide a default generic implementation for toJSON.

To use the second, simply add a deriving Generic clause to your datatype and declare a ToJSON instance for your datatype without giving definitions for toJSON or toEncoding.

For example, the previous example can be simplified to a more minimal instance:

{-# LANGUAGE DeriveGeneric #-}

import GHC.Generics

data Coord = Coord { x :: Double, y :: Double } deriving Generic

instance ToJSON Coord where
    toEncoding = genericToEncoding defaultOptions

Why do we provide an implementation for toEncoding here? The toEncoding function is a relatively new addition to this class. To allow users of older versions of this library to upgrade without having to edit all of their instances or encounter surprising incompatibilities, the default implementation of toEncoding uses toJSON. This produces correct results, but since it performs an intermediate conversion to a Value, it will be less efficient than directly emitting an Encoding. Our one-liner definition of toEncoding above bypasses the intermediate Value.

If DefaultSignatures doesn't give exactly the results you want, you can customize the generic encoding with only a tiny amount of effort, using genericToJSON and genericToEncoding with your preferred Options:

instance ToJSON Coord where
    toJSON     = genericToJSON defaultOptions
    toEncoding = genericToEncoding defaultOptions

Minimal complete definition

Nothing